[forms] move step forms in a specific directory
The directory is `src/components/stepforms` and its `index.ts` must
import all form components so that they registered automatically in the
`STEPFORM_REGISTRY`.

Note that we therefore still have to import `stepforms/` manually once at
some point.
