+++ src/components/Vqb.vue
@@ -23,11 +23,9 @@ import { Getter, Mutation, State } from 'vuex-class';
 import { VQBState } from '@/store/state';
 import { Pipeline, PipelineStep } from '@/lib/steps';
 import DataViewer from '@/components/DataViewer.vue';
-import DeleteColumnStepForm from '@/components/DeleteColumnStepForm.vue';
-import FillnaStepForm from '@/components/FillnaStepForm.vue';
-import RenameStepForm from '@/components/RenameStepForm.vue';
 import PipelineComponent from '@/components/Pipeline.vue';
 import ResizablePanels from '@/components/ResizablePanels.vue';
+import '@/components/stepforms';  // required to load all step forms
 import { STEPFORM_REGISTRY } from './formlib';
 
 import _ from 'lodash';
@@ -35,9 +33,6 @@ import _ from 'lodash';
 @Component({
   components: {
     DataViewer,
-    DeleteColumnStepForm,
-    FillnaStepForm,
-    RenameStepForm,
     Pipeline: PipelineComponent,
     ResizablePanels,
   },
