+++ src/components/stepforms/DeleteColumnStepForm.vue
@@ -1,17 +1,16 @@
 <template>
   <div>
     <div class="step-edit-form">
-      <h1 class="step-edit-form__title">FILL NULL VALUES STEP</h1>
+      <h1 class="step-edit-form__title">DELETE COLUMN STEP</h1>
     </div>
     <WidgetAutocomplete
       id="columnInput"
-      v-model="step.column"
-      name="Fill null values in:"
+      v-model="column"
+      name="Delete column:"
       :options="columnNames"
-      @input="setSelectedColumns({ column: step.column })"
+      @input="setSelectedColumns({ column })"
       placeholder="Enter a column"
     ></WidgetAutocomplete>
-    <WidgetInputText id="valueInput" v-model="step.value" name="With:" placeholder="Enter a value"></WidgetInputText>
     <div class="widget-form-action">
       <button
         class="widget-form-action__button widget-form-action__button--validate"
@@ -35,34 +34,31 @@ import _ from 'lodash';
 import { Mixins, Prop, Watch } from 'vue-property-decorator';
 import FormMixin from '@/mixins/FormMixin.vue';
 import { Pipeline } from '@/lib/steps';
-import fillnaSchema from '@/assets/schemas/fillna-step__schema.json';
-import WidgetInputText from './WidgetInputText.vue';
-import WidgetAutocomplete from '@/components/WidgetAutocomplete.vue';
+import deleteSchema from '@/assets/schemas/delete-column-step__schema.json';
+import WidgetAutocomplete from '@/components/stepforms/WidgetAutocomplete.vue';
 import { Getter, Mutation, State } from 'vuex-class';
 import { StepFormComponent } from '@/components/formlib';
+import { MutationCallbacks } from '@/store/mutations';
 
-interface FillnaStepConf {
-  column: string;
-  value: string;
+interface DeleteColumnStepConf {
+  columns: string[];
 }
 
 @StepFormComponent({
-  vqbstep: 'fillna',
-  name: 'fillna-step-form',
+  vqbstep: 'delete',
+  name: 'delete-step-form',
   components: {
     WidgetAutocomplete,
-    WidgetInputText,
   },
 })
-export default class FillnaStepForm extends Mixins(FormMixin) {
+export default class DeletStepForm extends Mixins(FormMixin) {
   @Prop({
     type: Object,
     default: () => ({
-      column: '',
-      value: '',
+      columns: [''],
     }),
   })
-  initialValue!: FillnaStepConf;
+  initialValue!: DeleteColumnStepConf;
 
   @Prop({
     type: Boolean,
@@ -70,13 +66,14 @@ export default class FillnaStepForm extends Mixins(FormMixin) {
   })
   isStepCreation!: boolean;
 
-  step: FillnaStepConf = { ...this.initialValue };
+  // Only manage the deletion of 1 column at once at this stage
+  column: string = this.initialValue.columns[0];
 
   @State pipeline!: Pipeline;
   @State selectedStepIndex!: number;
 
-  @Mutation selectStep!: (payload: { index: number }) => void;
-  @Mutation setSelectedColumns!: (payload: { column: string }) => void;
+  @Mutation selectStep!: MutationCallbacks['selectStep'];
+  @Mutation setSelectedColumns!: MutationCallbacks['setSelectedColumns'];
 
   @Getter selectedColumns!: string[];
   @Getter columnNames!: string[];
@@ -85,22 +82,22 @@ export default class FillnaStepForm extends Mixins(FormMixin) {
   @Watch('selectedColumns')
   onSelectedColumnsChanged(val: string[], oldVal: string[]) {
     if (!_.isEqual(val, oldVal)) {
-      this.step.column = val[0];
+      this.column = val[0];
     }
   }
 
   created() {
-    this.schema = fillnaSchema;
-    this.setSelectedColumns({ column: this.initialValue.column });
+    this.schema = deleteSchema;
+    this.setSelectedColumns({ column: this.initialValue.columns[0] });
   }
 
   validateStep() {
-    const ret = this.validator(this.step);
+    const ret = this.validator({ column: this.column });
     if (ret === false) {
       this.errors = this.validator.errors;
     } else {
       this.errors = null;
-      this.$emit('formSaved', { name: 'fillna', ...this.step });
+      this.$emit('formSaved', { name: 'delete', columns: [this.column] });
     }
   }
 
@@ -112,7 +109,7 @@ export default class FillnaStepForm extends Mixins(FormMixin) {
 }
 </script>
 <style lang="scss" scoped>
-@import '../styles/_variables';
+@import '../../styles/_variables';
 
 .widget-form-action__button {
   @extend %button-default;
