+++ src/components/stepforms/index.ts
@@ -1,0 +1,3 @@
+import './DeleteColumnStepForm.vue';
+import './FillnaStepForm.vue';
+import './RenameStepForm.vue';
